+++ app/(main)/location/components/MapSection/index.tsx
@@ -11,8 +11,8 @@ import VehicleDetailsCard from '@/app/(main)/location/components/VehicleDetailCa
 import VehicleMarker from '@/app/(main)/location/components/VehicleMarker'
 import Map from '@/components/domain/map/Map'
 import { useMapStatus } from '@/hooks/useCurrentMapStatus'
-// import { clusterService } from '@/lib/apis/cluster'
-import { clusterService } from '@/lib/apis/cluster'
+// import { clusterService } from '@/lib/apis'
+import { clusterService } from '@/lib/apis'
 import { LatLng } from '@/types/location'
 import { ClusterInfoModel, MapState } from '@/types/map'
 import { VehicleDetailModel, VehicleInfoModel } from '@/types/vehicle'
