+++ app/(main)/route/components/RouteSearchPanel/index.tsx
@@ -11,7 +11,7 @@ import Modal from '@/components/common/Modal'
 import { ModalMessageType } from '@/components/common/Modal/types'
 import { ZOOM_LEVEL } from '@/constants/map'
 import { useSearchVehicle } from '@/hooks/useSearchVehicle'
-import { routeService } from '@/lib/apis/route'
+import { routeService } from '@/lib/apis'
 import { formatISODateToKorean } from '@/lib/utils/date'
 import { normalizeCoordinate } from '@/lib/utils/normalize'
 import { validateDateSelection } from '@/lib/utils/validation'
