+++ lib/apis/interceptors.ts
@@ -1,8 +1,7 @@
 import { AxiosError, AxiosInstance, AxiosResponse, InternalAxiosRequestConfig } from 'axios'
 
 import { API_URL } from '@/constants/api'
-import { authService } from '@/lib/apis/auth'
-import { httpClient } from '@/lib/apis/client'
+import { httpClient, authService } from '@/lib/apis'
 import { useAuthStore } from '@/stores/useAuthStore'
 
 interface CustomRequestConfig extends InternalAxiosRequestConfig {
