+++ lib/apis/route.ts
@@ -1,32 +1,40 @@
 import { DateTime } from '@/app/(main)/route/types/date'
-import { httpClient } from '@/lib/apis/client'
+import { httpClient } from '@/lib/apis'
 import { formatToISODate } from '@/lib/utils/date'
+import { LiveRouteParams, RouteParams } from '@/types/map'
 
 export const routeService = {
+    // 차량 이동 경로 이력 조회
     getVehicleRoutesData: async (vehicleId: string, startDate: DateTime, endDate: DateTime, interval = 60) => {
-        const formattedStartDate = formatToISODate(startDate)
-        const formattedEndDate = formatToISODate(endDate)
-
-        // console.log(vehicleId, formattedStartDate, formattedEndDate, interval)
+        const params: RouteParams = {
+            startTime: formatToISODate(startDate),
+            endTime: formatToISODate(endDate),
+            interval,
+        }
 
         const response = await httpClient.get(`api/v1/vehicle/${vehicleId}/routes`, {
-            params: {
-                startTime: formattedStartDate,
-                endTime: formattedEndDate,
-                interval,
-            },
+            params,
         })
-        console.log(response.data.result)
 
-        return response.data.result
+        const { result } = response.data
+
+        return result
     },
-    getVehicleLiveRoutesData: async (vehicleId: string) => {
+    // 차량 실시간 이동 경로 조회
+    getVehicleLiveRoutes: async (vehicleId: string) => {
+        // const currentTime = new Date()
+        const params: LiveRouteParams = {
+            currentTime: '2025-01-27T21:23:00',
+            // currentTime,
+        }
+
         const response = await httpClient.get(`api/v1/vehicle/${vehicleId}/recent/routes`, {
-            params: {
-                currentTime: '2025-01-27T21:23:00',
-            },
+            params,
         })
-        console.log(response.data.result.routes)
+
+        const { result } = response.data
+
+        console.log(result)
 
         // return response.data.result
     },
